重构模块类关联关系

+++ aicode/src/main/java/com/aicode/project/ctrl/ProjectModelClassController.java
@@ -3,18 +3,20 @@
  */
 package com.aicode.project.ctrl;
 
+import com.aicode.core.entity.Page;
+import com.aicode.core.entity.R;
 import com.aicode.project.entity.ProjectModelClass;
 import com.aicode.project.service.ProjectModelClassService;
 import com.aicode.project.vo.ProjectModelClassPageVO;
-import com.aicode.project.vo.ProjectModelClassSaveVO;
 import com.aicode.project.vo.ProjectModelClassVO;
-import com.aicode.core.entity.Page;
-import com.aicode.core.entity.PageVO;
-import com.aicode.core.entity.R;
+import com.alibaba.druid.util.StringUtils;
 import com.alibaba.fastjson.JSON;
 import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import io.swagger.annotations.*;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiImplicitParam;
+import io.swagger.annotations.ApiImplicitParams;
+import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -38,66 +40,90 @@ public class ProjectModelClassController {
 
 
     /**
+     * 查询模块下的类详情信息
+     *
+     * @param id
+     * @return BeanRet
+     */
+    @ApiOperation(value = "查询模块下的类详情信息", notes = "查询模块下的类详情信息")
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "id", value = "", paramType = "query")
+    })
+    @GetMapping(value = "/loadById/{id}")
+    @ResponseBody
+    public R loadById(@PathVariable Long id) {
+        if (id == null) {
+            return R.failed("");
+        }
+
+        ProjectModelClass projectModelClass = projectModelClassService.getById(id);
+        log.info(JSON.toJSONString(projectModelClass));
+        return R.success(projectModelClass);
+    }
+
+    /**
      * 创建 模块下的类
      *
      * @return R
      */
     @ApiOperation(value = "创建ProjectModelClass", notes = "创建ProjectModelClass")
-    @PostMapping("/build")
-    public ProjectModelClassSaveVO build(@ApiParam(name = "创建ProjectModelClass", value = "传入json格式", required = true)
-                                   @RequestBody ProjectModelClassSaveVO projectModelClassSaveVO) {
-        if (null == projectModelClassSaveVO) {
-            return null;
-        }
-        ProjectModelClass newProjectModelClass = new ProjectModelClass();
-        BeanUtils.copyProperties(projectModelClassSaveVO, newProjectModelClass);
-
-        projectModelClassService.save(newProjectModelClass);
-
-        projectModelClassSaveVO = new ProjectModelClassSaveVO();
-        BeanUtils.copyProperties(newProjectModelClass, projectModelClassSaveVO);
-        log.debug(JSON.toJSONString(projectModelClassSaveVO));
-        return projectModelClassSaveVO;
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "id", value = "", paramType = "query", required = true),
+            @ApiImplicitParam(name = "mapClassTableCode", value = "类编码", paramType = "query", required = true),
+            @ApiImplicitParam(name = "projectModelCode", value = "模块编码", paramType = "query", required = true)
+    })
+    @PostMapping({"/build", "save"})
+    public ProjectModelClassVO build(@ApiIgnore ProjectModelClass projectModelClass) {
+        projectModelClassService.save(projectModelClass);
+        log.debug(JSON.toJSONString(projectModelClass));
+        return JSON.parseObject(JSON.toJSONString(projectModelClass), ProjectModelClassVO.class);
     }
 
 
     /**
-     * 根据条件mapClassTableCode查询模块下的类一个详情信息
+     * 查询模块下的类详情信息
      *
      * @param mapClassTableCode 类编码
-     * @return ProjectModelClassVO
+     * @return BeanRet
      */
-    @ApiOperation(value = "创建ProjectModelClass", notes = "创建ProjectModelClass")
-    @GetMapping("/load/mapClassTableCode/{mapClassTableCode}")
-    public ProjectModelClassVO loadByMapClassTableCode(@PathVariable java.lang.String mapClassTableCode) {
-        if (mapClassTableCode == null) {
-            return null;
+    @ApiOperation(value = "查询模块下的类详情信息", notes = "查询模块下的类详情信息")
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "mapClassTableCode", value = "类编码", paramType = "query")
+    })
+    @GetMapping(value = "/loadByMapClassTableCode")
+    @ResponseBody
+    public R loadByMapClassTableCode(@PathVariable String mapClassTableCode) {
+        if (StringUtils.isEmpty(mapClassTableCode)) {
+            return R.failed("");
         }
         ProjectModelClass projectModelClass = projectModelClassService.getOne(new LambdaQueryWrapper<ProjectModelClass>()
                 .eq(ProjectModelClass::getMapClassTableCode, mapClassTableCode));
-        ProjectModelClassVO projectModelClassVO = new ProjectModelClassVO();
-        BeanUtils.copyProperties(projectModelClass, projectModelClassVO);
-        log.debug(JSON.toJSONString(projectModelClassVO));
-        return projectModelClassVO;
+        log.info(JSON.toJSONString(projectModelClass));
+        return R.success(projectModelClass);
     }
+
+
     /**
-     * 根据条件projectModelCode查询模块下的类一个详情信息
+     * 查询模块下的类详情信息
      *
      * @param projectModelCode 模块编码
-     * @return ProjectModelClassVO
+     * @return BeanRet
      */
-    @ApiOperation(value = "创建ProjectModelClass", notes = "创建ProjectModelClass")
-    @GetMapping("/load/projectModelCode/{projectModelCode}")
-    public ProjectModelClassVO loadByProjectModelCode(@PathVariable java.lang.String projectModelCode) {
-        if (projectModelCode == null) {
-            return null;
+    @ApiOperation(value = "查询模块下的类详情信息", notes = "查询模块下的类详情信息")
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "projectModelCode", value = "模块编码", paramType = "query")
+    })
+    @GetMapping(value = "/loadByProjectModelCode")
+    @ResponseBody
+    public R loadByProjectModelCode(@PathVariable String projectModelCode) {
+        if (StringUtils.isEmpty(projectModelCode)) {
+            return R.failed("");
         }
+
         ProjectModelClass projectModelClass = projectModelClassService.getOne(new LambdaQueryWrapper<ProjectModelClass>()
                 .eq(ProjectModelClass::getProjectModelCode, projectModelCode));
-        ProjectModelClassVO projectModelClassVO = new ProjectModelClassVO();
-        BeanUtils.copyProperties(projectModelClass, projectModelClassVO);
-        log.debug(JSON.toJSONString(projectModelClassVO));
-        return projectModelClassVO;
+        log.info(JSON.toJSONString(projectModelClass));
+        return R.success(projectModelClass);
     }
 
     /**
@@ -109,20 +135,34 @@ public class ProjectModelClassController {
     @ApiImplicitParams({
             @ApiImplicitParam(name = "curPage", value = "当前页", required = true, paramType = "query"),
             @ApiImplicitParam(name = "pageSize", value = "分页大小", required = true, paramType = "query"),
+            @ApiImplicitParam(name = "id", value = "", paramType = "query"),
+            @ApiImplicitParam(name = "mapClassTableCode", value = "类编码", paramType = "query"),
+            @ApiImplicitParam(name = "projectModelCode", value = "模块编码", paramType = "query")
     })
     @GetMapping(value = "/list")
-    public PageVO<ProjectModelClassVO> list(@ApiIgnore ProjectModelClassPageVO projectModelClassVO, Integer curPage, Integer pageSize) {
+    public R list(@ApiIgnore ProjectModelClassPageVO projectModelClassVO, Integer curPage, Integer pageSize) {
         Page<ProjectModelClass> page = new Page<>(pageSize, curPage);
         QueryWrapper<ProjectModelClass> queryWrapper = new QueryWrapper<>();
+        if (projectModelClassVO.getId() > 0) {
+            queryWrapper.lambda().eq(ProjectModelClass::getId, projectModelClassVO.getId());
+        }
+
+        if (!StringUtils.isEmpty(projectModelClassVO.getMapClassTableCode())) {
+            queryWrapper.lambda().eq(ProjectModelClass::getMapClassTableCode, projectModelClassVO.getMapClassTableCode());
+        }
+
+        if (!StringUtils.isEmpty(projectModelClassVO.getProjectModelCode())) {
+            queryWrapper.lambda().eq(ProjectModelClass::getProjectModelCode, projectModelClassVO.getProjectModelCode());
+        }
+
         int total = projectModelClassService.count(queryWrapper);
-        PageVO<ProjectModelClassVO> projectModelClassVOPageVO = new PageVO<>();
         if (total > 0) {
             List<ProjectModelClass> projectModelClassList = projectModelClassService.list(queryWrapper, page.genRowStart(), page.getPageSize());
-            projectModelClassVOPageVO.setTotalRow(total);
-            projectModelClassVOPageVO.setRecords(JSON.parseArray(JSON.toJSONString(projectModelClassList),ProjectModelClassVO.class));
+            page.setTotalRow(total);
+            page.setRecords(projectModelClassList);
             log.debug(JSON.toJSONString(page));
         }
-        return projectModelClassVOPageVO;
+        return R.success(page);
     }
 
 
@@ -132,14 +172,15 @@ public class ProjectModelClassController {
      * @return R
      */
     @ApiOperation(value = "修改ProjectModelClass", notes = "修改ProjectModelClass")
+    @ApiImplicitParams({
+            @ApiImplicitParam(name = "id", value = "", paramType = "query"),
+            @ApiImplicitParam(name = "mapClassTableCode", value = "类编码", paramType = "query"),
+            @ApiImplicitParam(name = "projectModelCode", value = "模块编码", paramType = "query")
+    })
     @PutMapping("/modify")
-    public boolean modify(@ApiParam(name = "修改ProjectModelClass", value = "传入json格式", required = true)
-                          @RequestBody ProjectModelClassVO projectModelClassVO) {
-        ProjectModelClass newProjectModelClass = new ProjectModelClass();
-        BeanUtils.copyProperties(projectModelClassVO, newProjectModelClass);
-        boolean isUpdated = projectModelClassService.update(newProjectModelClass, new LambdaQueryWrapper<ProjectModelClass>()
-                .eq(ProjectModelClass::getId, projectModelClassVO.getId()));
-        return isUpdated;
+    public R modify(@ApiIgnore ProjectModelClass projectModelClass) {
+        projectModelClassService.updateById(projectModelClass);
+        return R.success();
     }
 
 
@@ -155,11 +196,20 @@ public class ProjectModelClassController {
             @ApiImplicitParam(name = "projectModelCode", value = "模块编码", paramType = "query")
     })
     @DeleteMapping("/delete")
-    public R delete(@ApiIgnore ProjectModelClassVO projectModelClassVO) {
-        ProjectModelClass newProjectModelClass = new ProjectModelClass();
-        BeanUtils.copyProperties(projectModelClassVO, newProjectModelClass);
+    public R delete(Long id, String mapClassTableCode, String projectModelCode) {
+        if (id == null) {
+            return R.failed("");
+        }
+        if (mapClassTableCode == null) {
+            return R.failed("");
+        }
+        if (projectModelCode == null) {
+            return R.failed("");
+        }
         projectModelClassService.remove(new LambdaQueryWrapper<ProjectModelClass>()
-                .eq(ProjectModelClass::getId, projectModelClassVO.getId()));
+                .eq(ProjectModelClass::getId, id)
+                .eq(ProjectModelClass::getProjectModelCode, projectModelCode)
+                .eq(ProjectModelClass::getMapClassTableCode, mapClassTableCode));
         return R.success("删除成功");
     }
